refactor(core,ui): VerboseOutputVisitore -> StructTreeTablePrinter
